+++ packages/account/test/bech32.test.ts
@@ -5,7 +5,9 @@ describe('bech32', () => {
 		const plaintext = 'Hello Radix!'
 		const bech32DataHex = '09011216181b030f04010906021903090f001010'
 		const bech32Data = Buffer.from(bech32DataHex, 'hex')
-		const decodedBech32Data = Bech32.convertDataFromBech32(bech32Data)._unsafeUnwrap()
+		const decodedBech32Data = Bech32.convertDataFromBech32(
+			bech32Data,
+		)._unsafeUnwrap()
 		expect(decodedBech32Data.toString('utf8')).toBe(plaintext)
 
 		const convertedToBech32Data = Bech32.convertDataToBech32(