+++ packages/application/src/dto/_types.ts
@@ -1,4 +1,9 @@
-import { AccountT, AddressT, ResourceIdentifierT, ValidatorAddressT } from '@radixdlt/account'
+import {
+	AccountT,
+	AddressT,
+	ResourceIdentifierT,
+	ValidatorAddressT,
+} from '@radixdlt/account'
 import {
 	ActionInput,
 	ActionType,
@@ -98,7 +103,7 @@ export type TransactionIntent = Readonly<{
 export type ValidatorsRequestInput = Readonly<{
 	size: number
 	// Address of last seen validator in list
-	cursor: string
+	cursor?: string
 }>
 
 export enum TransactionTrackingEventType {