+++ packages/application/src/dto/index.ts
@@ -2,5 +2,4 @@ export * from './_types'
 
 export * from './transactionIdentifier'
 export * from './transactionIntentBuilder'
-export * from '@radixdlt/account/dist/addresses/resourceIdentifier'
 export * from './tokenPermissions'