+++ packages/application/test/transferTokensAction.test.ts
@@ -1,65 +1,65 @@
-// import { Denomination } from '@radixdlt/primitives'
-// import { Amount, zero } from '@radixdlt/primitives/src/amount'
-// import { ResourceIdentifier, IntendedTransferTokens, TransferTokensInput, alice , bob } from '../src'
-//
-// describe.skip('TransferTokensActions', () => {
-//
-// 	const resourceIdentifier = ResourceIdentifier.create({
-// 		hash: alice.publicKey.asData({ compressed: true }),
-// 		name: 'FOOBAR',
-// 	})._unsafeUnwrap({ withStackTrace: true })
-// 	const amount = Amount.fromUnsafe(6, Denomination.Atto)._unsafeUnwrap({ withStackTrace: true })
-//
-// 	const input = <TransferTokensInput>{
-// 		to: bob,
-// 		amount: amount,
-// 		tokenIdentifier: resourceIdentifier,
-// 	}
-//
-// 	it(`should have a 'recipient' equal to 'input.to'.`, () => {
-// 		const tokenTransfer = IntendedTransferTokens.create(
-// 			input,
-// 			alice,
-// 		)._unsafeUnwrap({ withStackTrace: true })
-// 		expect(tokenTransfer.to.equals(bob)).toBe(true)
-// 	})
-//
-// 	it(`should have an 'amount' equal to 'input.amount'.`, () => {
-// 		const tokenTransfer = IntendedTransferTokens.create(
-// 			input,
-// 			alice,
-// 		)._unsafeUnwrap({ withStackTrace: true })
-// 		expect(tokenTransfer.amount.equals(amount)).toBe(true)
-// 	})
-//
-// 	it(`should have a 'sender' equal to 'input.from'.`, () => {
-// 		const tokenTransfer = IntendedTransferTokens.create(
-// 			input,
-// 			alice,
-// 		)._unsafeUnwrap({ withStackTrace: true })
-// 		expect(tokenTransfer.from.equals(alice)).toBe(true)
-// 	})
-//
-// 	it(`should have a 'uuid'`, () => {
-// 		const tokenTransfer = IntendedTransferTokens.create(
-// 			input,
-// 			alice,
-// 		)._unsafeUnwrap({ withStackTrace: true })
-// 		expect(tokenTransfer.uuid).toBeDefined()
-// 	})
-//
-// 	it('should be possible to transfer 0 tokens', () => {
-// 		const tokenTransfer = IntendedTransferTokens.create(
-// 			{ ...input, amount: zero },
-// 			alice,
-// 		)._unsafeUnwrap({ withStackTrace: true })
-//
-// 		expect(tokenTransfer.amount.equals(zero)).toBe(true)
-// 	})
-// })
+import { Address, ResourceIdentifier } from '@radixdlt/account'
+import { Denomination } from '@radixdlt/primitives'
+import { Amount, zero } from '@radixdlt/primitives/src/amount'
+import { TransferTokensInput } from '../src/actions/_types'
+import { IntendedTransferTokens } from '../src/actions/intendedTransferTokensAction'
 
-describe('empty', () => {
-	it('needs tests', () => {
+describe('TransferTokensActions', () => {
+ 	const resourceIdentifier = ResourceIdentifier.create({
+ 		hash: alice.publicKey.asData({ compressed: true }),
+ 		name: 'FOOBAR',
+ 	})._unsafeUnwrap({ withStackTrace: true })
 
+	const resourceIdentifier = ResourceIdentifier.fromPublicKeyAndName({
+		publicKey: alice.publicKey,
+		name: 'foobar',
+	})._unsafeUnwrap()
+	const amount = Amount.fromUnsafe(6, Denomination.Atto)._unsafeUnwrap()
+
+	const input = <TransferTokensInput>{
+		to: bob,
+		amount: amount,
+		tokenIdentifier: resourceIdentifier,
+	}
+
+	it(`should have a 'recipient' equal to 'input.to'.`, () => {
+		const tokenTransfer = IntendedTransferTokens.create(
+			input,
+			alice,
+		)._unsafeUnwrap()
+		expect(tokenTransfer.to.equals(bob)).toBe(true)
+	})
+
+	it(`should have an 'amount' equal to 'input.amount'.`, () => {
+		const tokenTransfer = IntendedTransferTokens.create(
+			input,
+			alice,
+		)._unsafeUnwrap()
+		expect(tokenTransfer.amount.equals(amount)).toBe(true)
+	})
+
+	it(`should have a 'sender' equal to 'input.from'.`, () => {
+		const tokenTransfer = IntendedTransferTokens.create(
+			input,
+			alice,
+		)._unsafeUnwrap()
+		expect(tokenTransfer.from.equals(alice)).toBe(true)
+	})
+
+	it(`should have a 'uuid'`, () => {
+		const tokenTransfer = IntendedTransferTokens.create(
+			input,
+			alice,
+		)._unsafeUnwrap()
+		expect(tokenTransfer.uuid).toBeDefined()
+	})
+
+	it('should be possible to transfer 0 tokens', () => {
+		const tokenTransfer = IntendedTransferTokens.create(
+			{ ...input, amount: zero },
+			alice,
+		)._unsafeUnwrap()
+
+		expect(tokenTransfer.amount.equals(zero)).toBe(true)
 	})
 })